Разбила скрипт на модули
